+++ appParsers/ParseSVG.py
@@ -447,7 +447,7 @@ def getsvggeo(node, object_type, root=None, units='MM', res=64, factor=1.0, app=
     else:
         log = logging.getLogger('base2')
 
-    kind = re.search('(?:\{.*\})?(.*)$', node.tag).group(1)
+    kind = re.search('(?:\{.*})?(.*)$', node.tag).group(1)
     geo = []
 
     # Recurse
@@ -558,7 +558,7 @@ def getsvgtext(node, object_type, app, units='MM'):
     :return:            List of Shapely geometry
     :rtype:             list
     """
-    kind = re.search('(?:\{.*\})?(.*)$', node.tag).group(1)
+    kind = re.search('(?:\{.*})?(.*)$', node.tag).group(1)
     geo = []
 
     # Recurse